+++ src/databases/CGNS/CGNS.xml
@@ -12,6 +12,14 @@
     <FilePatterns>
       *.cgns
     </FilePatterns>
+    <Files components="M">
+      avtCGNSFileReader.C
+      avtCGNSFileFormat.C
+    </Files>
+    <Files components="E">
+      avtCGNSFileReader.C
+      avtCGNSFileFormat.C
+    </Files>
     <Attribute name="" purpose="" persistent="true">
     </Attribute>
   </Plugin>
